workiing with VM
