+++ DiscontBankApp/DiscontBankApp.xcodeproj/xcuserdata/nivben-porath.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -64,8 +64,8 @@
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "52"
             endingLineNumber = "52"
-            landmarkName = "collectionView(_:cellForItemAt:)"
-            landmarkType = "7">
+            landmarkName = "unknown"
+            landmarkType = "0">
          </BreakpointContent>
       </BreakpointProxy>
    </Breakpoints>
